+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Improve the behavior of multivalueField component.
 - Update styles for "save page" component.
 - Update styles for progress bar component.
+- Update tooltip component.
 
 ### Deprecated
 -